Keep either url or response
I disagree with removing the response option, I'm not sure why I
accepted it. I think that it's nice to be able to use either a url or
response, it's helpful when the user wants to check something about the
response that akamai-rspec doesn't support. I also don't want to have to
change the existing tests without a good reason.

I think I may have just accepted it because I was so happy that a
stranger sent me a PR, and that clouded my judgement.